+++ R/argparse.R
@@ -258,7 +258,7 @@ parse_args_output <- function(output) {
 
 # @param argument argument to be converted from R to Python
 convert_argument <- function(argument, as_list = FALSE) {
-    if (is.character(argument)) argument <- paste0('"""', argument, '"""')
+    if (is.character(argument)) argument <- convert_character(argument)
     if (is.numeric(argument)) argument <- as.character(argument)
     if (is.logical(argument)) argument <- ifelse(argument, "True", "False")
     if (is.null(argument)) argument <- "None"
@@ -268,6 +268,11 @@ convert_argument <- function(argument, as_list = FALSE) {
         argument <- sprintf("(%s)", paste(argument, collapse = ", "))
     }
     argument
+}
+
+convert_character <- function(s) {
+    bool <- substr(s, nchar(s), nchar(s)) == '"'
+    ifelse(bool, paste0("'''", s, "'''"), paste0('"""', s, '"""'))
 }
 
 get_python_type <- function(type, proposed_arguments) {
@@ -364,9 +369,12 @@ assert_python_cmd <- function(python_cmd) {
     }
 }
 
-detects_python <- function() {
+required_modules <- c("argparse", "json | simplejson")
+
+detects_python <- function(...) {
     python_cmd <- try(find_python_cmd())
-    !inherits(python_cmd, "try-error")
+    !inherits(python_cmd, "try-error") &&
+        findpython::is_python_sufficient(python_cmd, ..., required_modules = required_modules)
 }
 
 # Internal function to find python cmd
@@ -376,7 +384,6 @@ find_python_cmd <- function(python_cmd = NULL) {
         python_cmd <- getOption("python_cmd")
     }
     if (is.null(python_cmd)) {
-        required_modules <- c("argparse", "json | simplejson")
         did_find_python3 <- findpython::can_find_python_cmd(minimum_version = "3.0",
                                                 required_modules = required_modules,
                                                 silent = TRUE)